added hooks
