+++ src/app/(with-sidebar)/featured/author/page.tsx
@@ -1,66 +1,13 @@
 "use client";
 import { Box, Grid2 as Grid, Typography } from "@mui/material";
-import React, { useCallback, useEffect, useState } from "react";
-import { trpc } from "~/server/client";
-import { type FeaturedAuthor } from "../../myfeed/types/types";
-import config from "~/app/_config/config";
 import Loader from "~/app/_components/loader/Loader";
 import UserNameProfile from "~/app/_components/userNameProfile/UserNameProfile";
 import Link from "next/link";
 import FollowButton from "~/app/_components/followButton/FollowButton";
-import { useClerk } from "@clerk/nextjs";
+import useFeaturedAuthorPage from "./_hooks/useFeaturedAuthorPage";
 
 const Page = () => {
-  const { user } = useClerk();
-  const [author, setAuthor] = useState<FeaturedAuthor[]>([]);
-  const [skip, setSkip] = useState(0);
-  const [hasMoreAuthor, setHasMoreAuthor] = useState<boolean | undefined>(true);
-
-  const featuredAuthorMutation = trpc.featuredAuthor;
-  const userMutation = trpc.user;
-
-  const { data: userFollowing } = userMutation.getUserFollowings.useQuery({
-    userEmail: user?.primaryEmailAddress?.emailAddress ?? "",
-  });
-
-  const { data, isLoading, error } =
-    featuredAuthorMutation.getFeaturedAuthors.useQuery(
-      {
-        skip,
-        limit:
-          skip === 0
-            ? config.lazyLoading.initialLimit
-            : config.lazyLoading.limit,
-      },
-      {
-        enabled: skip >= 0 && hasMoreAuthor === true,
-      },
-    );
-
-  const handleScroll = useCallback(() => {
-    const { scrollTop, scrollHeight, clientHeight } = document.documentElement;
-    const bottomReached = scrollHeight - scrollTop - clientHeight < 10;
-
-    if (bottomReached && !isLoading && !error) {
-      setHasMoreAuthor(data?.hasMoreAuthor);
-      setSkip((prev) =>
-        prev == 0
-          ? config.lazyLoading.initialLimit + prev
-          : prev + config.lazyLoading.limit,
-      );
-    }
-  }, [error, data?.hasMoreAuthor, isLoading]);
-
-  useEffect(() => {
-    setAuthor((prevAuthor) => [...prevAuthor, ...(data?.featuredAuthor ?? [])]);
-  }, [data]);
-
-  useEffect(() => {
-    window.addEventListener("scroll", handleScroll);
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, [handleScroll, hasMoreAuthor, isLoading]);
+  const { author, isLoading, userFollowing } = useFeaturedAuthorPage();
 
   return (
     <Box
@@ -78,6 +25,7 @@ const Page = () => {
           alignItems: "start",
           px: "4px",
           pt: "8px",
+          marginBottom: "10px",
           borderBottom: "1px solid #E0E0E0",
         }}
       >
